refactor(pms): move supplier contact contracts
